pyproject.toml
[tool.poetry]
name = "tiercepiscinebot"
version = "0.1.0"
description = ""
authors = ["TulipeSoyeuse <romain.dupeux24@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord-py = "^2.3.2"
python-dotenv = "^1.0.1"
requests-oauthlib = "^1.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
